refactor(idgen): rename package and struct according to Go conventions
Renamed the nodeid package to idgen/ and the nodeid.go file inside the package to idgen.go. Also renamed the nodeID struct to identifier and the NodeID interface to Identifier. The NewNodeID function was also renamed to NewIdentifier. The package of the test was also renamed accordingly. These changes were made to follow the Go conventions as suggested in the Effective Go documentation. Additionally, the package will now also be used to generate IDs for transactions.
