+++ pkg/idgen_test/idgen_test.go
@@ -1,39 +1,39 @@
 // Package nodeid contains tests for the nodeid package functionalities.
-package nodeid_test
+package idgen_test
 
 import (
 	"testing"
 
 	"github.com/google/uuid"
-	"github.com/pierreleocadie/SecuraChain/pkg/nodeid"
+	"github.com/pierreleocadie/SecuraChain/pkg/idgen"
 )
 
-// TestNewNodeID tests the NewNodeID function.
-// It ensures that the function returns a valid NodeID instance
+// TestNewIdentifier tests the NewIdentifier function.
+// It ensures that the function returns a valid Identifier instance
 // and that the identifier is a valid UUID.
-func TestNewNodeID(t *testing.T) {
+func TestNewIdentifier(t *testing.T) {
 	t.Parallel()
 
-	node := nodeid.NewNodeID()
+	node := idgen.NewIdentifier()
 
 	if node == nil {
-		t.Fatal("NewNodeID returned nil")
+		t.Fatal("NewIdentifier returned nil")
 	}
 
 	_, err := uuid.Parse(node.String())
 	if err != nil {
-		t.Fatalf("NewNodeID did not generate a valid UUID: %v", err)
+		t.Fatalf("NewIdentifier did not generate a valid UUID: %v", err)
 	}
 
 	t.Logf("The generated UUID identifier is: %s", node.String())
 }
 
-// TestNodeIDString tests the String method of the NodeID type.
+// TestIdentifierString tests the String method of the Identifier type.
 // It verifies that the method returns the correct string representation of the UUID.
-func TestNodeIDString(t *testing.T) {
+func TestIdentifierString(t *testing.T) {
 	t.Parallel()
 
-	node := nodeid.NewNodeID()
+	node := idgen.NewIdentifier()
 	_, err := uuid.Parse(node.String())
 	if err != nil {
 		t.Fatalf("String() did not return a valid UUID: %v", err)
@@ -42,16 +42,16 @@ func TestNodeIDString(t *testing.T) {
 	t.Logf("The generated UUID identifier is: %s", node.String())
 }
 
-// TestNodeIDUniqueness tests the uniqueness of the UUIDs generated by NewNodeID.
-// It generates a large number of NodeID instances and checks that there are no duplicates.
-func TestNodeIDUniqueness(t *testing.T) {
+// TestIdentifierUniqueness tests the uniqueness of the UUIDs generated by NewIdentifier.
+// It generates a large number of Identifier instances and checks that there are no duplicates.
+func TestIdentifierUniqueness(t *testing.T) {
 	t.Parallel()
 
 	iterations := 100000
 	ids := make(map[string]bool)
 
 	for i := 0; i < iterations; i++ {
-		node := nodeid.NewNodeID()
+		node := idgen.NewIdentifier()
 		if _, exists := ids[node.String()]; exists {
 			t.Fatalf("Duplicate UUID generated: %s", node.String())
 		}
